server/requirements.txt
django
psycopg2-binary
gunicorn
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
daphne
channels-redis
celery
redis
openai
numpy
scikit-learn
requests
python-dotenv